refactor(money): migrate monetary values to BigDecimal

+++ src/main/java/com/ecommerce/project/repositories/OrderRepository.java
@@ -6,13 +6,14 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
 public interface OrderRepository extends JpaRepository<Order, Long> {
 
     @Query("SELECT COALESCE(SUM(o.totalAmount),0) FROM Order o")
-    Double getTotalRevenue();
+    BigDecimal getTotalRevenue();
 
     List<Order> findByEmailOrderByOrderDateDesc(String email);
